Stylistic changes (PEP8, packaging)
This commit also introduces some changes to make
the package run under python3, even though it is still not
working properly
