+++ scoary.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
-# Scoary - Microbial Pan-GWAS. Associates genes in Roary output with phenotypes.
+# Scoary - Microbial Pan-GWAS. Associates genes in Roary output with phenotypes
 # Copyright (C) 2016  Ola Brynildsrud (ola.brynildsrud@fhi.no)
 
 # This program is free software: you can redistribute it and/or modify
@@ -17,16 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import sys
-
-if (sys.version_info >= (3,0)):
-	sys.stdout.write("You seem to be using Python 3. Scoary was written for Python 2.7.x. Try 'python2.7 Scoary.py [arguments]'\n")
-	sys.exit(1)
-elif (sys.version_info < (2,7)):
-	sys.stdout.write("Warning: Scoary was written for Python 2.7.x. Your version is " + str(sys.version) + "\n")
-	sys.stdout.write("Please make sure to use Python 2.7 to avoid errors.\n")
-	
-import Scoary_methods
+from scoary import methods
 
 if __name__ == '__main__':
-	Scoary_methods.main()
+    methods.main()
